Refactor all webpack to use a simples config, remove all react, babel, eslint, mocha, chai and sinon dependencies
