+++ index.html
@@ -5,14 +5,14 @@
 		<meta name="description" content="You app for smart urban transport">
 		<meta name="keywords" content="Transport, ride, urban, car, share">
 		<meta name="author" content="Hege Refsnes">
-		<link rel="stylesheet" type="text/css" href="app.css">
+		<link rel="stylesheet" type="text/css" href="main.css">
 	</head>
 	<body>
 		<div id="page-wrap">
 			<div id="header">
 				<div id="topbar">
 					<div id="title">
-					<img src="https://s3-sa-east-1.amazonaws.com/llago-bucket/caronaboard_logo.png" style="float:left; padding: 10px;"/>
+					<img src="images/caronaboard_logo.png" style="float:left; padding: 10px;"/>
 					<h1>Carona Board</h1>
 					<h2>Seu carro não precisa levar apenas você</h2>
 					</div>
