+++ package.json
@@ -4,14 +4,12 @@
   "description": "Social ride network",
   "main": "index.js",
   "scripts": {
-    "start": "NODE_ENV=production node ./src/main-server",
+    "start": "node src/server.js",
     "postinstall": "npm run build",
-    "dev-server": "NODE_ENV=development node ./src/main-server",
-    "build": "npm run clean && webpack --config ./webpack/webpack.config.js",
-    "clean": "rm -rf ./build",
-    "lint": "eslint ./src --ext .js --ext .jsx --cache",
-    "test": "NODE_ENV=test mocha --compilers js:babel-core/register,.:mocha-image-compiler --recursive --check-leaks --full-trace ./src/**/*-spec.js",
-    "ci": "npm run lint && npm run test"
+    "dev-server": "webpack-dev-server",
+    "build": "webpack -p",
+    "test": "echo 'not implemented yet ):'",
+    "ci": "npm run test"
   },
   "repository": {
     "type": "git",
@@ -24,54 +22,23 @@
   },
   "homepage": "https://github.com/CaronaBoard/caronaboard#readme",
   "dependencies": {
-    "babel-core": "^6.7.7",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-0": "^6.5.0",
     "body-parser": "^1.15.0",
     "bootstrap": "^3.3.6",
-    "chalk": "^1.1.3",
-    "css-loader": "^0.23.1",
+    "copy-webpack-plugin": "^3.0.1",
+    "css-loader": "^0.25.0",
     "elm-webpack-loader": "^3.0.6",
-    "eslint": "^2.9.0",
-    "eslint-loader": "^1.3.0",
-    "eslint-plugin-react": "^5.0.1",
     "express": "^4.13.4",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
     "html-webpack-plugin": "^2.16.0",
-    "jquery": "^2.2.3",
-    "marked": "^0.3.5",
     "node-sass": "^3.7.0",
-    "purifycss-webpack-plugin": "^2.0.3",
-    "react": "^15.0.1",
-    "react-bootstrap": "^0.29.4",
-    "react-dom": "^15.0.1",
-    "react-elm-components": "^1.0.0",
-    "sass-loader": "^3.2.0",
+    "sass-loader": "^4.0.2",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
     "webpack": "^1.13.0",
-    "webpack-build-notifier": "^0.1.8",
-    "webpack-merge": "^0.12.0",
     "webpack-validator": "^1.1.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.7.7",
-    "babel-preset-react-hmre": "^1.1.1",
-    "chai": "^3.5.0",
-    "chai-enzyme": "^0.4.2",
-    "enzyme": "^2.3.0",
-    "expect": "^1.20.1",
-    "jsdom": "^9.0.0",
-    "mocha": "^2.4.5",
-    "mocha-image-compiler": "^1.0.0",
-    "react-addons-test-utils": "^15.0.2",
-    "react-shallow-testutils": "^2.0.0",
-    "sinon": "^1.17.4",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-dev-server": "^1.14.1"
+    "webpack-dev-server": "^1.16.2"
   }
 }
