--- webpack/config/webpack.production.config.js
@@ -1,38 +1,0 @@
-const webpack = require('webpack');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const WebpackBuildNotifierPlugin = require('webpack-build-notifier');
-const PurifyCssPlugin = require("purifycss-webpack-plugin");
-const PATHS = require('../PATHS');
-
-const config = {
-  module: {
-    loaders: [
-      {
-        test: /\.scss$/,
-        loader: ExtractTextPlugin.extract('style-loader', 'css-loader?sourceMap!sass-loader?sourceMap')
-      },
-      {
-        test: /\.css$/,
-        loader: ExtractTextPlugin.extract('style-loader', 'css-loader?sourceMap!sass-loader?sourceMap')
-      }
-    ]
-  },
-  plugins: [
-    new webpack.DefinePlugin({
-      'process.env.NODE_ENV': '"production"',
-      'process.env.BABEL_ENV': '"production"'
-    }),
-    new webpack.optimize.UglifyJsPlugin({
-      compress: {
-        warnings: false
-      }
-    }),
-    new ExtractTextPlugin('[name].css'),
-    new PurifyCssPlugin({
-      basePath: __dirname,
-      paths: ['../../**/*.html']
-    })
-  ]
-}
-
-module.exports = config;
